meet-it/requirements.txt
# Web framework
fastapi==0.109.1
uvicorn[standard]==0.23.2
pydantic==2.6.0
python-multipart==0.0.6

# TensorFlow + dependencies
tensorflow==2.19.1
numpy==2.1.3
h5py==3.15.1

# Vision + ML utilities
opencv-python==4.10.0.84
mediapipe==0.10.14
Pillow==10.1.0

# Optional for extra utilities (if needed)
# scikit-learn==1.5.2
# pandas==2.2.3
# matplotlib==3.9.2
